+++ tests/compile_fail/try_from/duplicate_empty_type.rs
@@ -1,0 +1,8 @@
+#[derive(derive_more::TryFrom)]
+#[try_from(usize)]
+enum Enum {
+    Field,
+    Field2,
+}
+
+fn main() {}