+++ tests/compile_fail/try_from/empty_type.rs
@@ -1,0 +1,7 @@
+#[derive(derive_more::TryFrom)]
+#[try_from(usize)]
+enum Enum {
+    Field,
+}
+
+fn main() {}