+++ tests/try_from.rs
@@ -85,3 +85,188 @@ fn test_discriminants_on_enum_with_fields() {
     assert!(Enum::try_from(-14).is_err());
     assert_eq!(Enum::EmptyTuple(), Enum::try_from(-13).unwrap());
 }
+
+#[test]
+fn test_try_from_types() {
+    #[derive(Debug, PartialEq, Eq)]
+    struct F1;
+
+    impl TryFrom<usize> for F1 {
+        type Error = ();
+
+        fn try_from(value: usize) -> Result<Self, Self::Error> {
+            if value == 1 {
+                return Ok(Self);
+            }
+            Err(())
+        }
+    }
+
+    #[derive(Debug, PartialEq, Eq)]
+    struct F2;
+
+    impl TryFrom<usize> for F2 {
+        type Error = ();
+
+        fn try_from(value: usize) -> Result<Self, Self::Error> {
+            if value == 2 {
+                return Ok(Self);
+            }
+            Err(())
+        }
+    }
+
+    #[derive(TryFrom, Debug, PartialEq, Eq)]
+    #[try_from(usize)]
+    enum Enum {
+        Field(F1),
+        Field2 { x: F2 },
+    }
+
+    assert_eq!(Enum::Field(F1), Enum::try_from(1).unwrap());
+    assert_eq!(Enum::Field2 { x: F2 }, Enum::try_from(2).unwrap());
+    assert!(Enum::try_from(3).is_err());
+}
+
+#[test]
+fn test_try_from_types_custom_unit_error() {
+    #[derive(Debug, PartialEq, Eq)]
+    struct Error;
+
+    #[derive(Debug, PartialEq, Eq)]
+    struct F1;
+
+    impl TryFrom<usize> for F1 {
+        type Error = Error;
+
+        fn try_from(value: usize) -> Result<Self, Self::Error> {
+            if value == 1 {
+                return Ok(Self);
+            }
+            Err(Error)
+        }
+    }
+
+    #[derive(Debug, PartialEq, Eq)]
+    struct F2;
+
+    impl TryFrom<usize> for F2 {
+        type Error = Error;
+
+        fn try_from(value: usize) -> Result<Self, Self::Error> {
+            if value == 2 {
+                return Ok(Self);
+            }
+            Err(Error)
+        }
+    }
+
+    #[derive(TryFrom, Debug, PartialEq, Eq)]
+    #[try_from(usize, Error)]
+    enum Enum {
+        Field(F1),
+        Field2 { x: F2 },
+    }
+
+    assert_eq!(Enum::Field(F1), Enum::try_from(1).unwrap());
+    assert_eq!(Enum::Field2 { x: F2 }, Enum::try_from(2).unwrap());
+    assert_eq!(Err(Error), Enum::try_from(3));
+}
+
+#[test]
+fn test_try_from_types_custom_error() {
+    #[derive(Debug, PartialEq, Eq)]
+    enum Error {
+        FromEnum,
+        FromVariant,
+    }
+
+    #[derive(Debug, PartialEq, Eq)]
+    struct F1;
+
+    impl TryFrom<usize> for F1 {
+        type Error = Error;
+
+        fn try_from(value: usize) -> Result<Self, Self::Error> {
+            if value == 1 {
+                return Ok(Self);
+            }
+            Err(Error::FromVariant)
+        }
+    }
+
+    #[derive(Debug, PartialEq, Eq)]
+    struct F2;
+
+    impl TryFrom<usize> for F2 {
+        type Error = Error;
+
+        fn try_from(value: usize) -> Result<Self, Self::Error> {
+            if value == 2 {
+                return Ok(Self);
+            }
+            Err(Error::FromVariant)
+        }
+    }
+
+    assert_eq!(Err(Error::FromVariant), F2::try_from(3));
+
+    #[derive(TryFrom, Debug, PartialEq, Eq)]
+    #[try_from(usize, Error, Error::FromEnum)]
+    enum Enum {
+        Field(F1),
+        Field2 { x: F2 },
+    }
+
+    assert_eq!(Enum::Field(F1), Enum::try_from(1).unwrap());
+    assert_eq!(Enum::Field2 { x: F2 }, Enum::try_from(2).unwrap());
+    assert_eq!(Err(Error::FromEnum), Enum::try_from(3));
+}
+
+#[test]
+fn test_try_from_multiple_types_custom_error() {
+    #[derive(Debug, PartialEq, Eq)]
+    enum Error {
+        FromEnum,
+        FromVariant,
+    }
+
+    #[derive(Debug, PartialEq, Eq)]
+    struct F1;
+
+    impl TryFrom<usize> for F1 {
+        type Error = Error;
+
+        fn try_from(value: usize) -> Result<Self, Self::Error> {
+            if value == 1 {
+                return Ok(Self);
+            }
+            Err(Error::FromVariant)
+        }
+    }
+
+    #[derive(Debug, PartialEq, Eq)]
+    struct F2;
+
+    impl TryFrom<usize> for F2 {
+        type Error = Error;
+
+        fn try_from(value: usize) -> Result<Self, Self::Error> {
+            if value == 1 {
+                return Ok(Self);
+            }
+            Err(Error::FromVariant)
+        }
+    }
+
+    assert_eq!(Err(Error::FromVariant), F2::try_from(3));
+
+    #[derive(TryFrom, Debug, PartialEq, Eq)]
+    #[try_from(usize, Error, Error::FromEnum)]
+    enum Enum {
+        Field(F1, F2),
+    }
+
+    assert_eq!(Enum::Field(F1, F2), Enum::try_from(1).unwrap());
+    assert_eq!(Err(Error::FromEnum), Enum::try_from(2));
+}